feat: Add OAuth token validation middleware to API group
Introduce `ValidateToken` middleware to secure API routes. This middleware extracts the bearer token, validates it against the Users service (with caching for performance), and sets the authenticated user in the request.

It ensures that only requests with valid and active tokens can access protected API endpoints by prepending it to the `api` middleware group.

Add the `auth:api` middleware to user routes